requirements.txt
pylint
coverage
lxml
cssutils
numpy
cython
pyfakefs
mox3
portalocker
toposort
pyaudio
vext
vext.pyqt5
asynctest
quamash
./3rdparty/csound/
posix_ipc
eventfd
psutil
./3rdparty/lv2/
./3rdparty/lilv/
intervaltree
pkgconfig
pycapnp
